Create Library management system
This mini project in C Library Management System is a console application without graphic developed using the C programming language. It is compiled in Code::Blocks with the GCC compiler. In this, you can perform book-related operations like in a REAL library management system with computer
For library management, this project considers six departments – Computer, Electrical, Civil, Electronics, Mechanical and Architecture. These departments work simultaneously with the operations mentioned above. You can add a book to the Civil section, delete a book from the Electrical section or view issued book details of the Mechanical department.

These are the functions used in this mini project:

void mainmenu(void) – This function is used to display the main menu of this project. Scroll down to Output to view the main menu.

void returnfunc(void) – Inside this function, the main menu function (mentioned above) is called when the user presses a key. With this, the user can return back to the main menu.

void addbooks(void) – This function adds books in a file. For that, you need to mention the department to which you want to add the book. The record is saved in a file. And, it is similar for the following functions as well.

void deletebooks(void)
void editbooks(void)
void searchbooks(void)
void issuebooks(void)
void viewbooks(void)

void issuerecord() – With this function, you can keep record of the student to whom the book has been issued.

void closeapplication(void) – This function is for closing the application.

int  getdata() – This function asks for data input from the user.

int checkid(int) – This function is used to check whether the ID of a book entered by a user exists in file or not.

void Password() – Due to this function, the user is asked to input password to run the application after it is opened. You can’t change the password upon running the application.

void gotoxy (int x, int y) – You need to understand this function as it is one of the most important one used in mini project on Library Management System in C. This function allows you to print text in any place of screen. Using this function in Code::Blocks requires coding, but it can be directly used in Turbo C. Here is a code for this function in Code::Blocks.

COORD coord = {0, 0};  // sets coordinates to (0,0) as global variables
void gotoxy (int x, int y)
{
        coord.X = x; coord.Y = y; // X and Y are the coordinates
        SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), coord);
}
delay(unsigned int mseconds) – This function delays the execution. It can be used directly in Turbo C, but requires coding in Code::Blocks. The code is given below:

 void delay(unsigned int mseconds)
{
    clock_t goal = mseconds + clock();
    while (goal > clock());
}Here, you can perform functions such as add books, return books, issue books, delete record of books issued, view record of books issued, search for books, and more. File handling has been extensively used in this project for almost all functions. So, this project can definitely guide you to understand C mini projects in a better way.

The source code is organized well, and it has multiple comment lines to help you understand the project better. The whole code is around 1000 lines, so I haven’t displayed it hereThis project is complete and totally error-free. Understanding this long but simple project will definitely help you create similar projects in C programming (such as bank management, school billing, hospital management etc.) and solve problems by compiling and debugging.

So, try understanding and analyzing the source code of this C mini project on library management system. If you have any suggestions or queries regarding this project, do comment them belo